Stop reward from printing stuff, make models never delete plexon files (in case the matching was bad)

+++ db/tracker/models.py
@@ -320,5 +320,6 @@ class DataFile(models.Model):
         return dict(system=self.system.name, path=self.path)
 
     def remove(self, **kwargs):
-        os.unlink(self.path)
+        if (system != System.objects.get(name="plexon"))
+            os.unlink(self.path)
         super(DataFile, self).remove(**kwargs)
